+++ src/tui/popup.rs
@@ -53,6 +53,7 @@ impl Popup {
       ("Space".to_string(), "Toggle Timer".to_string()),
       ("n".to_string(), "Next Session".to_string()),
       ("r".to_string(), "Reset Session".to_string()),
+      ("a".to_string(), "To add 5 minutes".to_string()),
       ("q".to_string(), "Quit".to_string()),
       ("]".to_string(), "Next page".to_string()),
       ("[".to_string(), "Previous page".to_string()),